// src/api/prefect.tsx
export interface components {
    schemas: {
// ...
        Body_count_flow_runs_flow_runs_count_post: {
            flows?: components["schemas"]["FlowFilter"];
            flow_runs?: components["schemas"]["FlowRunFilter"];
            task_runs?: components["schemas"]["TaskRunFilter"];
            deployments?: components["schemas"]["DeploymentFilter"];
            work_pools?: components["schemas"]["WorkPoolFilter"];
            work_pool_queues?: components["schemas"]["WorkQueueFilter"];
        };
// ...
        Body_read_deployments_deployments_filter_post: {
            /**
             * Offset
             * @default 0
             */
            offset: number;
            flows?: components["schemas"]["FlowFilter"];
            flow_runs?: components["schemas"]["FlowRunFilter"];
            task_runs?: components["schemas"]["TaskRunFilter"];
            deployments?: components["schemas"]["DeploymentFilter"];
            work_pools?: components["schemas"]["WorkPoolFilter"];
            work_pool_queues?: components["schemas"]["WorkQueueFilter"];
            /** @default NAME_ASC */
            sort: components["schemas"]["DeploymentSort"];
            /**
             * Limit
             * @description Defaults to PREFECT_API_DEFAULT_LIMIT if not provided.
             */
            limit?: number;
        };
// ...
        Body_read_flow_runs_flow_runs_filter_post: {
            /** @default ID_DESC */
            sort: components["schemas"]["FlowRunSort"];
            /**
             * Offset
             * @default 0
             */
            offset: number;
            flows?: components["schemas"]["FlowFilter"] | null;
            flow_runs?: components["schemas"]["FlowRunFilter"] | null;
            task_runs?: components["schemas"]["TaskRunFilter"] | null;
            deployments?: components["schemas"]["DeploymentFilter"] | null;
            work_pools?: components["schemas"]["WorkPoolFilter"] | null;
            work_pool_queues?: components["schemas"]["WorkQueueFilter"] | null;
            /**
             * Limit
             * @description Defaults to PREFECT_API_DEFAULT_LIMIT if not provided.
             */
            limit?: number;
        };
// ...
        Flow: {
            /**
             * Id
             * Format: uuid
             */
            id?: string;
            /** Created */
            created?: string | null;
            /** Updated */
            updated?: string | null;
            /**
             * Name
             * @description The name of the flow
             */
            name: string;
            /**
             * Tags
             * @description A list of flow tags
             */
            tags?: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}